+++ app/ThirdParty/ThirdPartySupport.php
@@ -92,7 +92,9 @@ class ThirdPartySupport {
 	 */
 	private function get_third_party_support(): array {
 		$list = array(
+			'ExternalFilesInMediaLibrary\ThirdParty\BlockEditor',
 			'ExternalFilesInMediaLibrary\ThirdParty\Downloadlist',
+			'ExternalFilesInMediaLibrary\ThirdParty\Elementor',
 			'ExternalFilesInMediaLibrary\ThirdParty\EnableMediaReplace',
 			'ExternalFilesInMediaLibrary\ThirdParty\Exmage',
 			'ExternalFilesInMediaLibrary\ThirdParty\Massedge',