+++ changelog.md
@@ -6,6 +6,7 @@
 
 - Added support for image meta-data (like caption) for external hosted images
 - Added more hooks
+- Added Third Party support for Block Editor and Elementor
 
 ### Optimized